A few changes to wobble and also advancements in the screen and scoring system.
Score is now half working. the game never ends but it keeps track of your score. The screen now displays the drink name and the score. The bottlecap opening looks a lot better now. Going to try and round out the gameplay loop by adding a button to start playing and also fix the shader issues on the glass possibly. Productive week so far, want to keep this momentum!