Added buttons and also instantiation machine for colas
Added pressable buttons, still need work but will get on that soon. Added in a drink dispensing machine that uses the buttons to give the player a coke. Going to add more buttons for tonics and etc. Think its a better system than filling the bar with loads of mixers as it would get clumsy and also its better if they have an infinite amount.